Todo

+++ app.py
@@ -105,7 +105,8 @@ def logout():
    return redirect ('/')
 #VER CLIENTE
 @app.route('/cliente/<id>', methods=['GET'])
-def get_disp(id):     
+def get_disp(id):
+  if "user" in session:       
     desired_value = id
     response = supabase.table('clientes').select('*').eq('id', desired_value).limit(1).execute()
     data = response.data
@@ -116,10 +117,12 @@ def get_disp(id):
     if not data:
       return "Not Found", 404
     return render_template('disp_cliente.html', data=data, timestamp=formatted_timestamp, alias =alias)
+  return redirect ('/')
 
 ##AÑADIR CLIENTE
 @app.route ('/addcliente', methods = ['POST', 'GET'])
 def add_cliente():
+ if "user" in session: 
    nombre = request.form.get ('nombre')
    Alias = request.form.get ('Alias')
    rfc = request.form.get ('rfc')
@@ -160,10 +163,12 @@ def add_cliente():
           'tecnicoTelefono' : tecnicoTelefono    
           }).execute()
       return redirect('/dashboard')
+ return redirect ('/')
    
 ##EDITAR CLIENTE
 @app.route ('/cliente/edit/<id>', methods = ['POST', 'GET'])
 def cliente_edit (id):
+  if "user" in session:
    formid = request.form.get('form_id')
    print (formid)
    if formid == 'form1':
@@ -264,29 +269,35 @@ def cliente_edit (id):
          return "Not Found", 404
       else:
          return render_template ('/disp_cliente.html', data=data, timestamp=formatted_timestamp )
+  return redirect ('/')
       
 
 ## BORRAR CLIENTE
 @app.route('/cliente/delete/<id>', methods = ['POST', 'GET'])
 def disp_delete (id):
+  if "user" in session: 
    response = supabase.table('clientes').delete().eq('id', id).execute()
    if not response:
       return "Not Found", 404
    else:
       return redirect ('/dashboard')
+  return redirect ('/') 
 
 ## VER DISPOSITIVOS DE CLIENTE 
 @app.route ('/cliente/dispositivos/<id>', methods = ['POST', 'GET'])
 def cliente_dispositivos (id):
+ if "user" in session:  
    data = supabase.table ('clientes').select('nombre').eq('id', id).execute()
    cliente = data.data[0]['nombre']
    response = supabase.table ('dispositivos').select('*').eq('cliente_id', id).execute()
    data = response.data
    return render_template ('/dispositivos.html', data = data, cliente = cliente, clienteid = id)
+ return redirect ('/')
 
 ## AÑADIR DISPOSITIVOS DE CLIENTE
 @app.route ('/cliente/dispositivos/add/<id>', methods = ['POST', 'GET'])
 def add_dispositivos (id):
+ if "user" in session:  
    data = supabase.table ('clientes').select('nombre').eq('id', id).execute()
    cliente = data.data[0]['nombre']
    MAC_data = request.form.get('MAC')
@@ -306,10 +317,12 @@ def add_dispositivos (id):
           'cliente_id' : Cliente_id   
           }).execute()
    return redirect (url_for('cliente_dispositivos', id  = Cliente_id)) 
+ return redirect ("/")
 
 ##VER DISPOSITIVO ESPECíFICO   
 @app.route ('/cliente/dispositivos/ver/<clienteid>/<id>', methods = ['POST', 'GET'])
 def ver_dispositivo (clienteid, id):
+ if 'user' in session:  
    response = supabase.table ('clientes').select('nombre').eq('id', clienteid).execute()
    cliente = response.data[0]['nombre']
    dispositivo_id = id
@@ -319,19 +332,23 @@ def ver_dispositivo (clienteid, id):
    timestampz_obj = parser.isoparse(timestampz)
    formatted_timestamp = timestampz_obj.strftime("%Y-%m-%d %H:%M:%S %Z")
    return render_template ('/ver_dispositivo.html', cliente = cliente,clienteid = clienteid , data = data, timestamp=formatted_timestamp)
+ return redirect ('/')
 
 ##BORRAR DISPOSITIVO
 @app.route ('/cliente/dispositivos/ver/<clienteid>/delete/<id>', methods = ['POST', 'GET'])
 def delete_dispositivo (clienteid, id):
+ if 'user' in session:  
    response = supabase.table('dispositivos').delete().eq('id', id).execute()
    if not response:
       return "Not Found", 404
    else:
       return redirect (url_for('cliente_dispositivos', id  = clienteid)) 
+ return redirect ('/')  
    
 ##EDITAR DISPOSITIVO
 @app.route ('/cliente/dispositivos/ver/<clienteid>/edit/<id>', methods = ['POST', 'GET'])
 def edit_dispositivo (clienteid, id):
+  if 'user' in session:  
     Location_data= request.form.get('Location')
     Status_data = request.form.get ('Status')
     response = (
@@ -343,7 +360,8 @@ def edit_dispositivo (clienteid, id):
     .eq("id", id)
     .execute())
     return redirect (url_for('ver_dispositivo', clienteid = clienteid, id = id))
-   
+  return redirect ('/')
+ 
 ##VER USUARIOS
 @app.route ('/verusuarios', methods = ['POST', 'GET'])
 def ver_usuarios ():
@@ -354,10 +372,12 @@ def ver_usuarios ():
          data_clientes = response.data
          id = session ['user_cliente_id']
          return render_template ('ver_usuarios.html', data = data, data_clientes = data_clientes)
+      return redirect ('/')
       
 #VER USUARIO
 @app.route('/usuario/<id>', methods=['GET'])
-def get_usuario(id):     
+def get_usuario(id):
+  if 'user' in session:       
     desired_value = id
     response = supabase.table('usuarios').select('*').eq('id', desired_value).limit(1).execute()
     data = response.data
@@ -373,10 +393,11 @@ def get_usuario(id):
     if not data:
       return "Not Found", 404
     return render_template('disp_usuario.html', data=data, timestamp=formatted_timestamp, cliente = cliente, id = id, clienteid = clienteid)
-
+  return redirect ('/')
 ##EDITAR USUARIO
 @app.route ('/usuario/editar/<id>/<clienteid>', methods = ['POST', 'GET'])
 def editar_usuario(id, clienteid):
+ if "user" in session:  
    username = request.form.get ('Username')
    password = request.form.get ('Password')
    password = generate_password_hash (password)
@@ -405,6 +426,139 @@ def editar_usuario(id, clienteid):
       .eq("id", id)
       .execute())
    return redirect (url_for('get_usuario', id = id ))
+ return redirect ('/')
+
+##VER USUARIOS DE CLIENTE
+@app.route ('/cliente/verusuarioscliente/<id>', methods = ['POST', 'GET'])
+def ver_usuarios_cliente (id):
+ if "user" in session:  
+   data = supabase.table ('clientes').select('nombre').eq('id', id).execute()
+   cliente = data.data[0]['nombre']
+   response = supabase.table ('usuarios').select('*').eq('user_cliente_id', id).execute()
+   data = response.data
+   return render_template ('/ver_usuarios_cliente.html', data = data, cliente = cliente, clienteid = id)
+ return redirect ("/")
+
+##VER USUARIO ESPECÍFICO DE CLIENTE
+@app.route('/cliente/verusuarioscliente/usuario/<id>', methods=['GET'])
+def get_usuario_cliente(id):
+  if "user" in session:       
+    desired_value = id
+    response = supabase.table('usuarios').select('*').eq('id', desired_value).limit(1).execute()
+    data = response.data
+    clienteid = data [0]['user_cliente_id']
+    if clienteid == 0:
+       cliente = 'NA'
+    else:
+      response = supabase.table('clientes').select('nombre').eq('id', clienteid).execute()
+      cliente = response.data[0]['nombre']
+    timestampz = data[0]['created_at']
+    timestampz_obj = parser.isoparse(timestampz)
+    formatted_timestamp = timestampz_obj.strftime("%Y-%m-%d %H:%M:%S %Z")
+    if not data:
+      return "Not Found", 404
+    return render_template('disp_usuario_cliente.html', data=data, timestamp=formatted_timestamp, cliente = cliente, id = id, clienteid = clienteid)
+  return redirect ("/")
+
+##EDITAR USUARIO ESPECÍFICO DE CLIENTE
+@app.route ('/cliente/verusuarioscliente/usuario/edit/<id>/<clienteid>', methods = ['POST', 'GET'])
+def editar_usuario_cliente(id, clienteid):
+ if "user" in session:  
+   username = request.form.get ('Username')
+   password = request.form.get ('Password')
+   password = generate_password_hash (password)
+   role = request.form.get ('Role')
+   if role == "admin":
+      user_cliente_id = 0
+   else:
+      user_cliente_id = clienteid
+   print (user_cliente_id)
+   nombre = request.form.get ('Nombre')
+   apellido = request.form.get ('Apellido')
+   email = request.form.get ('Email')
+   telefono = request.form.get ('Telefono')
+   response = (
+      supabase.table("usuarios")
+      .update({
+          'username' : username,
+          'password' : password,
+          'role' : role,
+          'user_cliente_id' : user_cliente_id,
+          'Nombre' : nombre,
+          'Apellido' :  apellido,
+          'Email' : email,
+          'Telefono' : telefono
+         })
+      .eq("id", id)
+      .execute())
+   return redirect (url_for('get_usuario_cliente', id = id ))
+ return redirect ("/")
+
+##ELIMINAR USUARIO ESPECÍFICO DE CLIENTE
+@app.route('/cliente/verusuarioscliente/usuario/delete/usuario/<id>/<clienteid>' , methods = ['POST', 'GET'])
+def delete_usuario_cliente (id, clienteid):
+ if "user" in session:  
+   source_page = request.form.get('source_page')
+   if source_page == 'pagina_de_cliente':
+      response = supabase.table('usuarios').delete().eq('id', id).execute()
+      if not response:
+         return "Not Found", 404
+      else:
+         return redirect (url_for('ver_usuarios_cliente', id  = clienteid))
+   if source_page == 'pagina_de_admin':
+      response = supabase.table('usuarios').delete().eq('id', id).execute()
+      if not response:
+         return "Not Found", 404
+      else:
+         return redirect (url_for('verusuarios'))
+ return redirect ("/")     
+
+
+@app.route('/usuario/delete/<id>' , methods = ['POST', 'GET'])
+def delete_usuario (id):
+ if "user" in session:  
+   source_page = request.form.get('source_page')
+   if source_page == 'pagina_de_admin':
+      response = supabase.table('usuarios').delete().eq('id', id).execute()
+      if not response:
+         return "Not Found", 404
+      else:
+         return redirect (url_for('ver_usuarios'))
+ return redirect ("/")     
+      
+## AÑADIR USUARIO DE CLIENTE
+@app.route ('/addusuario/<id>', methods = ['POST', 'GET'])
+def add_usuario (id):
+ if "user" in session:  
+   source_page = request.form.get ('source_page')
+   data = supabase.table ('clientes').select('nombre').eq('id', id).execute()
+   cliente = data.data[0]['nombre']
+   if source_page == 'pagina_de_cliente':
+         Nombre = request.form.get('Nombre')
+         if not Nombre:
+            return  render_template('/add_usuario_cliente.html', cliente = cliente, clienteid = id)
+         else:
+            Apellido = request.form.get('Apellido')
+            Email = request.form.get ('Email')
+            Username = request.form.get ('Username')
+            Password = request.form.get ('Password')
+            Password = generate_password_hash (Password)
+            Telefono = request.form.get ('Telefono')
+            Role = request.form.get ('Role')
+            response = supabase.table('usuarios').insert({
+            'username': Username,
+            'password': Password,
+            'role' : Role,
+            'user_cliente_id' : id,
+            'Nombre' : Nombre,
+            'Apellido' : Apellido,
+            'Email' : Email,
+            'Telefono' : Telefono
+          }).execute()
+            return redirect (url_for ('ver_usuarios_cliente', id = id))
+   return render_template('/add_usuario_cliente.html', cliente = cliente, clienteid = id)
+ return redirect ("/")
+
 
 if __name__ == "__main__":
   app.run(host='0.0.0.0',debug=True)
